# public/openapi.yml
openapi: 3.0.4
info:
  title: Contract-Model-Model-View (CMMV)
  description: >-
    CMMV is a minimalist Node.js framework focused on contract-driven
    development, combining automatic code generation, RPC communication, and
    declarative programming to build efficient, scalable applications with
    simplified backend and frontend integration.
  version: 0.8.31
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    OAuth2:
      name: CMMV OAuth2
      type: oauth2
      scheme: bearer
      bearerFormat: JWT
      flows:
        password:
          authorizationUrl: /oauth2/authorize
          tokenUrl: /oauth2/token
          scopes:
            email: email
            openid: openid
            profile: profile
  schemas:
...
    Groups:
      properties:
        id:
          readOnly: true
          type: string
        name:
          type: string
          required: true
        roles:
          type: array
          default: []
          required: false
          items:
            type: string
      required:
        - name
    User:
      properties:
        id:
          readOnly: true
          type: string
        username:
          type: string
          required: true
        password:
          type: string
          required: true
        provider:
          type: string
          readOnly: true
          required: false
        groups:
          type: array
          readOnly: true
          required: false
          items:
            $ref: '#/components/schemas/Groups'
        roles:
          type: array
          readOnly: true
          default: []
          required: false
          items:
            type: string
        blocked:
          type: boolean
          readOnly: true
          required: true
          default: false
        validated:
          type: boolean
          readOnly: true
          required: true
          default: false
        verifyEmail:
          type: boolean
          readOnly: true
          required: true
          default: false
        verifySMS:
          type: boolean
          readOnly: true
          required: true
          default: false
        profile:
          type: string
          default: '{}'
          required: false
      required:
        - username
        - password
        - blocked
        - validated
        - verifyEmail
        - verifySMS
...
    Catch:
      properties:
        status:
          readOnly: true
          type: number
          default: 500
          required: true
        processingTime:
          readOnly: true
          type: number
          required: true
        message:
          readOnly: true
          type: string
          required: true
        requestId:
          readOnly: true
          type: string
          required: false
        telemetry:
          readOnly: true
          type: array
          required: false
      required:
        - status
        - processingTime
paths:
  /auth/login:
    post:
      tags:
        - Autorization
      summary: >-
        Route to login using username and password, requires release and depends
        on authorization on the server
      externalDocs:
        url: https://cmmv.io/docs/modules/authentication#login
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  required: true
                password:
                  type: string
                  required: true
                token:
                  type: string
                  required: false
                opt:
                  type: string
                  required: false
            examples:
              default:
                value:
                  username: ''
                  password: ''
              reCAPTCHA:
                value:
                  username: ''
                  password: ''
                  token: ''
              F2A:
                value:
                  username: ''
                  password: ''
                  token: ''
                  opt: ''
      responses:
        '200':
          description: Login response schema
          headers:
            set-cookie:
              schema:
                type: string
                description: Session cookie for authentication if application is enabled
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                  processingTime:
                    type: number
                  result:
                    type: object
                    properties:
                      token:
                        type: string
                      refreshToken:
                        type: string
        '500':
          description: 'Internal Server Error: An unexpected error occurred on the server.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Catch'
      parameters: []
      security: []
  /auth/register:
    post:
      tags:
        - Autorization
      summary: >-
        Route to register a new public user, requires release depends on
        authorization on the server
      externalDocs:
        url: https://cmmv.io/docs/modules/authentication#user-registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Registry response schema
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                  processingTime:
                    type: number
                  result:
                    type: object
                    properties:
                      message:
                        type: string
        '500':
          description: 'Internal Server Error: An unexpected error occurred on the server.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Catch'
      parameters: []
      security: []
  /auth/check:
    get:
      tags:
        - Autorization
      summary: Check if authentication is still valid
      externalDocs:
        url: null
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  sucess:
                    type: boolean
        '401':
          description: >-
            Unauthorized: The client must authenticate itself to get the
            requested response.
          content:
            application/json:
              schema:
                type: string
                example: Unauthorized
        '403':
          description: 'Forbidden: The client does not have access rights to the content.'
          content:
            application/json:
              schema:
                type: string
                example: Forbidden
        '500':
          description: 'Internal Server Error: An unexpected error occurred on the server.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Catch'
      parameters: []
      security:
        - BearerAuth: []
  /auth/check-username:
    post:
      tags:
        - Autorization
      summary: Checks if the user is already in use
      externalDocs:
        url: null
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
      responses:
        '200':
          content:
            text/plain:
              schema:
                type: boolean
        '500':
          description: 'Internal Server Error: An unexpected error occurred on the server.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Catch'
      parameters: []
      security: []
  /auth/refresh:
    post:
      tags:
        - Autorization
      summary: Route to refresh authentication token using refresh token
      externalDocs:
        url: https://cmmv.io/docs/modules/authentication#refresh-token
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                  processingTime:
                    type: number
                  result:
                    type: object
                    properties:
                      token:
                        type: string
        '500':
          description: 'Internal Server Error: An unexpected error occurred on the server.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Catch'
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
        - in: header
          name: refreshToken
          required: false
          schema:
            type: string
        - in: cookie
          name: refreshToken
          required: false
          schema:
            type: string
      security: []
...
  /user:
    get:
      tags:
        - User
      summary: Returns User records by filter
      externalDocs:
        url: null
      responses:
        '200':
          description: User response schema
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                  processingTime:
                    type: number
                  result:
                    type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
                      count:
                        type: number
                      pagination:
                        type: object
                        properties:
                          limit:
                            type: number
                          offset:
                            type: number
                          sortBy:
                            type: string
                            default: id
                          sort:
                            type: string
                            default: asc
                          search:
                            type: string
                          searchField:
                            type: string
                          filters:
                            type: object
        '401':
          description: >-
            Unauthorized: The client must authenticate itself to get the
            requested response.
          content:
            application/json:
              schema:
                type: string
                example: Unauthorized
        '403':
          description: 'Forbidden: The client does not have access rights to the content.'
          content:
            application/json:
              schema:
                type: string
                example: Forbidden
        '500':
          description: 'Internal Server Error: An unexpected error occurred on the server.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Catch'
      parameters: *ref_1
      security:
        - BearerAuth: []
    post:
      tags:
        - User
      summary: Insert new User
      externalDocs:
        url: null
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  nullable: false
                  uniqueItems: true
                  validations: &ref_2
                    - type: IsString
                      message: Invalid username
                    - type: MinLength
                      value: 4
                      message: Invalid username
                    - type: MaxLength
                      value: 40
                      message: Invalid username
                password:
                  type: string
                  nullable: false
                  validations: &ref_3
                    - type: IsString
                      message: Invalid password
                profile:
                  type: string
                  nullable: true
                  default: '''{}'''
      responses:
        '401':
          description: >-
            Unauthorized: The client must authenticate itself to get the
            requested response.
          content:
            application/json:
              schema:
                type: string
                example: Unauthorized
        '403':
          description: 'Forbidden: The client does not have access rights to the content.'
          content:
            application/json:
              schema:
                type: string
                example: Forbidden
        '500':
          description: 'Internal Server Error: An unexpected error occurred on the server.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Catch'
      parameters: []
      security:
        - BearerAuth: []
...
security: []